requirements.txt
fastapi
pydantic
uvicorn
python-multipart
python-decouple
langchain
python-dotenv
pinecone-client